+++ src/providers/auth.provider.tsx
@@ -2,6 +2,8 @@ import { AuthAction, authReducer, AuthState } from '@context/auth/auth.reducer';
 import { AuthContext, initialAuthState } from '@context/auth/auth.context';
 import React from 'react';
 import { authStorage } from '@context/auth/auth.storage';
+import jwt from 'jsonwebtoken';
+import { setUserData } from '@context/auth/auth.action-creators';
 
 interface AuthProviderProps {
   children: React.ReactNode;
@@ -18,6 +20,12 @@ export const AuthProvider = ({ children }: AuthProviderProps) => {
   React.useEffect(() => {
     authStorage.setAccessToken(state.accessToken);
     authStorage.setRefreshToken(state.refreshToken);
+
+    const jwtPayload = jwt.decode(state.accessToken) as { userId: string; username: string };
+
+    if (jwtPayload) {
+      dispatch(setUserData(jwtPayload.username, jwtPayload.userId));
+    }
   }, [state.accessToken, state.refreshToken]);
 
   return (